+++ browser-extras/src/main/java/org/realityforge/arez/browser/extras/spy/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * This package contains additional browser based utilities and abstractions based on the core Arez spy infrastructure.
+ */
+@Unsupported( "These utilities are based on unsupported core.spy package and are expected to evolve as needs dictate" )
+package org.realityforge.arez.browser.extras.spy;
+
+import org.realityforge.arez.Unsupported;